Set up quiz difficulty

+++ QuizZing/src/main/java/com/UndefinedParameter/app/core/Quiz.java
@@ -173,7 +173,7 @@ public class Quiz {
 	}
 	
 	@JsonProperty
-	public void setDifficulty(int difficulty) {
+	public void setDifficulty(float difficulty) {
 		this.difficulty = difficulty;
 	}
